Update CamCAN example scripts.
