+++ examples/camcan/1_preprocess.py
@@ -5,24 +5,28 @@
 import pathlib
 from glob import glob
 from dask.distributed import Client
+
 from osl import preprocessing, utils
 
+# Directories
 BASE_DIR = "/well/woolrich/projects/camcan"
 RAW_DIR = BASE_DIR + "/cc700/meg/pipeline/release005/BIDSsep/derivatives_rest/aa/AA_movecomp/aamod_meg_maxfilt_00002"
-PREPROC_DIR = BASE_DIR + "/winter23/preproc"
+PREPROC_DIR = BASE_DIR + "/summer23/preproc"
 
+# Settings
 config = """
     preproc:
     - crop: {tmin: 30}
-    - filter: {l_freq: 0.5, h_freq: 125, method: iir, iir_params: {order: 5, ftype: butter}}
-    - notch_filter: {freqs: 50 100}
-    - resample: {sfreq: 250}
+    - filter: {l_freq: 0.5, h_freq: 200, method: iir, iir_params: {order: 5, ftype: butter}}
+    - notch_filter: {freqs: 50 88 100 118 150 156 185, notch_widths: 2}
+    - resample: {sfreq: 400}
     - bad_segments: {segment_len: 500, picks: mag, significance_level: 0.1}
     - bad_segments: {segment_len: 500, picks: grad, significance_level: 0.1}
     - bad_segments: {segment_len: 500, picks: mag, mode: diff, significance_level: 0.1}
     - bad_segments: {segment_len: 500, picks: grad, mode: diff, significance_level: 0.1}
-    - bad_channels: {picks: meg, significance_level: 0.1}
-    - ica_raw: {picks: meg, n_components: 64}
+    - bad_channels: {picks: mag, significance_level: 0.1}
+    - bad_channels: {picks: grad, significance_level: 0.1}
+    - ica_raw: {picks: meg, n_components: 0.99}
     - ica_autoreject: {picks: meg, ecgmethod: correlation, eogthreshold: auto}
     - interpolate_bads: {}
 """
@@ -32,7 +36,7 @@ if __name__ == "__main__":
 
     # Get input files
     inputs = []
-    for subject in glob(RAW_DIR + "/sub-*"):
+    for subject in sorted(glob(RAW_DIR + "/sub-*")):
         subject = pathlib.Path(subject).stem
         inputs.append(RAW_DIR + f"/{subject}/mf2pt2_{subject}_ses-rest_task-rest_meg.fif")
